net/mlx5: Expose command polling interface
Add a new interface for commands execution that allows the
user to wait for the command's completion in a busy-wait
loop (polling mode).

This useful if we want to execute a command in a polling mode
while the driver is working in events mode for the rest of
the commands.

Signed-off-by: Majd Dibbiny <majd@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>