net/mlx5e: Add support to neighbour update flow
In order to offload TC encap rules the driver must find the IP tunnel
neighbour according to the output device and the destination IP given by the user.

In case the neighbour isn't valid, we won't be able to offload the flow
into the HW. We cache the flow in the driver in order to offload the rule when
the neighbour will be resolved.

When the flow is only cached in the driver and not offloaded into HW
yet, we use EAGAIN return value to mark this case. This return value is
relevant only inside the driver, to the upper layers the driver returns
success.

Listen on kernel neighbour update netevents to trace relevant neighbours
validity state:

1. If a neighbour becomes valid, offload the related rules to HW.
2. If the neighbour becomes invalid, remove the related rules from HW.
3. If the neighbour mac address was changed, update the encap header.
   Remove all the offloaded rules using the old encap header from the HW and
   insert new rules to HW with updated encap header.

A relevant neighbour is a neighbour used by one of the TC rules
offloaded to our driver. In order to keep tracking those neighbours
validity state, we hold the neighbours information (pair of device pointer and
destination IP) in an hash table on the egress representor and search
for a match each time we get neighbour update netevent.

Issue: 932305
Change-Id: Ic848aaed9b1ea31d02bd3bc1c047f4a12acc8189
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>