net/mlx5e: Use flag to properly monitor a flow rule offloading state
Instead of relaying on the 'flow->rule' pointer value which can be
valid or invalid (in case the FW returns an error while trying to offload
the rule), monitor the rule state using a flag.

In downstream patch which adding support to IP tunneling neigh upadet
flow, A TC rule could be cached in the driver and not offloaded into the
HW. In this case, the flow handle pointer stays NULL.
Add offloaded flag check to avoid crash when querying rule
statistics.

This patch doesn't add any new functionality.

Issue: 932305
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>
Change-Id: Ic37e386543ef4247724fa24486908edd1140ec03