net/mlx5e: Change the TC offload rule add/del code path to be per NIC or E-Switch
Refactor the code to deal with add/del TC rules to have handler per NIC/E-switch
offloading use case, and push the latter into the e-switch code. This provides
better separation and is be used in down-stream patches for a fix and for
features which apply only to e-switch or nic.

Issue: 840526
Change-Id: Ie0b736cba689f35e6a2ca4ae6f85d6f4f9ae74ae
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Reviewed-by: Roi Dayan <roid@mellanox.com>