net/mlx5e: Update neighbour 'used' state using HW flow rules counters
When IP tunnel encapsulation rules are offloaded, the kernel can't see
the traffic of the offloaded flow. The neighbour for the IP tunnel
destination of the offloaded flow can mistakenly become STALE and
deleted by the kernel since its 'used' value wasn't changed.

To make sure that a neighbour which is used by the HW won't become
STALE, we proactively update the neighbour 'used' value every
DELAY_PROBE_TIME period, when packets were matched and counted by the HW
for one of the tunnel encap flows related to this neighbour.

The periodic task that updates the used neighbours, is scheduled when a
tunnel encap rule is successfully offloaded into HW and keeps
re-scheduling itself as long as the representor's neighbours list isn't
empty.

Add, remove, lookup and status change operations done over the
representor's neighbours list or the neighbour hash entry encaps list
are all serialized by RTNL lock.

Issue: 932305
Change-Id: I575c25aa7253f12abf4a9f925c28d9dddd40c587
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>