net/mlx5e: Update neigh 'used' state using HW flow rules counters
When IP tunnel encapsulation rules are offloaded, the kernel can't see
the traffic of the offloaded flow. The neighbour for the IP tunnel
destination of the offloaded flow can mistakenly become STALE and
deleted by the kernel since its 'used' value wasn't changed.

To make sure that a neighbour which is used by the HW won't become
STALE, we proactively update the neighbour 'used' value every
DELAY_PROBE_TIME period, when packets were matched on one of the flows
related to this neigh.

The periodic task that update the used neighs, is scheduled when an
encap rule is successfully offloaded into HW and keeps re-scheduling itself
as long as the representor's neigh list isn't empty.

rtnl lock is taken in the neighbour statistics work function to protect
the representor neighbours list from removal operations.

Issue: 932305
Change-Id: I575c25aa7253f12abf4a9f925c28d9dddd40c587
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>