net/mlx5: Add gid_set function
Add gid_set function which programs a GID entry to the hardware.
This is required by a later patch that implements reserved-gids,
and its use in Innova FPGA code.

New file is added because we are going to extend it with reserved
GIDs mechanism in a later patch in the patchset.

Another patch in the patchset modifies mlx5_ib to use this
new utility function.

Issue: 1005200
Change-Id: I75ac49441c970d57a91b961824aad9294d0bad67
Signed-off-by: Ilan Tayari <ilant@mellanox.com>