IB/mlx5: Split MR allocation and population
mlx5 compatible devices (Connect-IB, ConnectX4/5/6) have two ways
of populating the MTT table of an MKEY: by issuing a FW command,
and by issuing UMR WQE.

A UMR is much faster, and it is preferable way for mlx5 operations,
so it should be used whenever possible.

Current code uses UMR only if the MKEY was allocated from the MR cache.
Let's fix the code to use UMR even for MKEYs that were allocated using
a FW command.

Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>