net/mlx5: Add Software-parser hardware descriptor layout
Software-Parser (swp) is a hardware feature in ConnectX which allows the
host software to specify protocol header offsets in the TX path, thus
overriding the hardware parser.
This is useful for protocols that the ASIC may not be able to parse on
its own.

Add software parser descriptors to the hardware structs.

Issue: 1005200
Change-Id: Ib253f54c035a193cb0c4c35a811e6c6b229a2ea3
Signed-off-by: Ilan Tayari <ilant@mellanox.com>
Reviewed-by: Boris Pismenny <borisp@mellanox.com>