net/mlx5e: IPSec, Add offload GSO support
Add GSO support for IPSec crypto-offload code path.

The FPGA needs to update the ESP header (seqno and seqiv) on the
resulting packets, so it needs to calculate the packet number within
the GSO back from the TCP sequence number.
To do that fast, the driver provides the value 1/gso_size in the
metadata.
Note that for GSO SKBs, the stack does not include an ESP trailer,
unlike the non-GSO case.

Issue: 1005200
Change-Id: I02bf3e569b70aa8799d5144929f8aab8b9a4c59c
Signed-off-by: Ilan Tayari <ilant@mellanox.com>
Signed-off-by: Yossi Kuperman <yossiku@mellanox.com>
Signed-off-by: Yevgeny Kliteynik <kliteyn@mellanox.com>
Reviewed-by: Boris Pismenny <borisp@mellanox.com>