net/mlx4_en: Add dynamic variable to hold the number of user priorities (UP)
Until this patch, the number of UPs was hard coded for eight.
Replace this with a variable in struct "mlx4_en_port_profile".
Currently, the variable will hold the maximum number of UP,
as before.
The patch creates an infrastructure to add an option for dynamic
change of the actual number of TCs.

issue: 1028830
Change-Id: I4de1780206ffd68341372361586905858825f8a6
Signed-off-by: Inbar Karmy <inbark@mellanox.com>