net/mlx5e: Fix slow xmit not working for ConnectX-4
The orignal commit only changed the code path for PF/VF, but it didn't
change the code path for VF representor.  And because mlx5e_priv is
initialized to 0, params->tx_min_inline_mode for VF representor is 0
as well.  But for ConnectX-4, it should be 1 to be working.

Issue: 1040416
Change-Id: Ib51577d73224fb0612185d5ab0836fe6ad723e65
Fixes: a6f402e49901 ("net/mlx5e: Tx, no inline copy on ConnectX-5")
Signed-off-by: Chris Mi <chrism@mellanox.com>