net/mlx5: Fix interface flags before cleanup in unload_one
In load_one, the interface flags are changed from down to up,
only after initializing the interfaces.
In unload_one, the flags are changed from up to down before the
interface cleanup.

Change the cleanup order to be opposite to initialization order.

This fixes flag consistency between init and cleanup.

Issue: 1005200
Change-Id: I763718e51559868f67cc0a76789fc70f39566006
Signed-off-by: Ilan Tayari <ilant@mellanox.com>