net/mlx5e: Disallow TC offloading of unsupported match/action combinations
When offloading header re-write, the HW may need to adjust checksums along
the packet. For IP traffic, and a case where we are asked to modify fields in
the IP header, current HW supports that only for TCP and UDP. Enforce it, in
this case fail the offloading attempt for non TCP/UDP packets.

issue: 1050641
Change-Id: Ia575dfd157bb965bcde03a79e1a95a633074bf38
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>