net/mlx5: Make get_cqe routine not ethernet-specific
Move mlx5e_get_cqe routine to wq.h and rename it to
mlx5_cqwq_get_cqe.

This allows it to be used by other CQ users outside of the
ethernet driver code.

A later patch in this patchset will make use of it from
FPGA code for the FPGA high-speed connection.

Issue: 1005200
Change-Id: Id10674ef8c9c68edd4ac078c31cebc3aa7483218
Signed-off-by: Ilan Tayari <ilant@mellanox.com>