net/mlx5e: Support offloading TC NIC hairpin flows
Hairpin is TC **NIC** rule that involves forwarding.

All hairpin rules from the current NIC device (called "func" in
the code) to a given peer are set on the same hairpin RQ/SQ pair.

The hairpin RQ/SQ pair is set on demand and removed when there
are no TC rules that need it. When we set the func --> peer pair
we immediately set also the peer --> func pair. We delete both when
there's no demand for either of them (note add/del are under rtnl).

TC rule that matches on ICMP, does header re-write of the dst mac and hairpin
from RX of P1 to TX on P2 (P1 and P2 are two VF devices in this example):

tc filter add dev enp1s2f1 protocol ip parent ffff: prio 2
  flower skip_sw ip_proto icmp
  action pedit ex munge eth dst set 10:22:33:44:55:66 pipe
  action mirred egress redirect dev enp1s2f2

issue: 1068457
Change-Id: I1ea222342c3189c27cdc0c0158828a870d5e558a
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>