net/mlx5e: Stop NAPI when irq balancer changes affinity
NAPI context keeps rescheduling on same CPU as long as it's busy.
This doesn't give the oppurtunity for changes in irq affinities
to take effect.
Fix that by calling napi_complete_done() upon a change in affinity.
This would stop the NAPI and reschedule it on the new CPU.

Issue: 1041183
Change-Id: Ia66aa21ede75782185ae6bbf7cd32cff20ba15e6
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>