net/mlx5e: Serialize PF vport disablement with netdev TC offload calls
Moving from legacy to switchdev mode and vise-versa through devlink, the
eswitch vports are disabled and then the esw steering tables are re-created.

The PF netdevice remains up and running throughout all this process,
as such it's possible that an offloaded TC call through the PF ndo
would be racing with the devlink call.

To prevent that, we serialize the PF vport state change with the eswitch
internal state lock and check the port state and the switch mode when
we're in the TC offload ndo.

Issue: 1013092
Change-Id: I00a560b3d451a005838a73637240fc39764e1090
Signed-off-by: Shahar Klein <shahark@mellanox.com>