IB/core: Introduce counter set object and its create/destroy verbs
Counter set instance is allocated on an IB context and belongs to
that context.
Upon successful creation the counter set hardware resources are
allocated and are ready to be bound to its counted type IB object.
Post the bind operation the hardware related counters could be queried.

Downstream patches in this series will introduce the bind and the
query functionality.

Counter set instance can be de-allocated, upon successful
destruction the related hardware resources are released.

Prior to calling the destroy verb the user must first make sure that
the counter set is not being used by its IB object, in other words not
bound to any of its counted objects.

Signed-off-by: Raed Salem <raeds@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>