RDMA/mthca: Make explicit conversion to 64bit value
The "lg" variable is declared as int so in all places there this variable
is used as a shift operand, the output will be int too.

This produces the following smatch warning:
drivers/infiniband/hw/mthca/mthca_cmd.c:701 mthca_map_cmd() warn:
	should '1 << lg' be a 64 bit type?

Simple declaration of "1" to be "1ULL" will fix the issue.

Signed-off-by: Leon Romanovsky <leonro@mellanox.com>