net/mlx5e: Remove unnecessary fields in ICO SQ
As of current design, in each NAPI, only a single UMR WQE
completion could be available in the completion queue of the
the internal control operations (ICO) send queue, in addition
to nop operations that require no actions upon completion.
This renders the consume index obsolete, as the wqe_counter
field in CQE is sufficient.

This helps removing a memory barrier, and obsoletes the need
for tracking the num_wqebbs to update the consumer counter.

In addition, remove other unused fields in icosq struct:
pdev, dma_fifo_pc, and prev_cc.

Issue: 1080498
Change-Id: Iad9178512cb5fe575b0377d3ec4009f458aeafd3
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>